Crate holochain_core_types

Holochain Core Types

The reason for having this crate is to have a minimal but complete set of types that are used in most other Holochain crates, but that don't include Holochain itself.

Note: This is already quite big. Maybe break the CAS and EAV traits out into their separate crate as well since those are generic and not necessarily bound to Holochain.

Modules

agent
Represents an agent entry in the cas

cas

chain_header

chain_migrate

crud_status

dna
dna is a library for working with holochain dna files/entries.

eav

entry

error

file_validation

hash

json

keys

link

signature

time

validation