Type Definition holochain_core_types::cas::content::Content

type Content = JsonString;

the Content is a JsonString this is the only way to be confident in persisting all Rust types across all backends