Trait holochain_core_types::cas::storage::ContentAddressableStorage

pub trait ContentAddressableStorage: Clone + Send + Sync + Debug {
    fn add(
        &mut self, 
        content: &AddressableContent
    ) -> Result<(), HolochainError>;

    fn contains(&self, address: &Address) -> Result<bool, HolochainError>;

    fn fetch(
        &self, 
        address: &Address
    ) -> Result<Option<Content>, HolochainError>;

    fn get_id(&self) -> Uuid;
}

content addressable store (CAS) implements storage in memory or persistently anything implementing AddressableContent can be added and fetched by address CAS is append only

Required Methods

fn add(&mut self, content: &AddressableContent) -> Result<(), HolochainError>

adds AddressableContent to the ContentAddressableStorage by its Address as Content

fn contains(&self, address: &Address) -> Result<bool, HolochainError>

true if the Address is in the Store, false otherwise. may be more efficient than retrieve depending on the implementation.

fn fetch(&self, address: &Address) -> Result<Option<Content>, HolochainError>

returns Some AddressableContent if it is in the Store, else None AddressableContent::from_content() can be used to allow the compiler to infer the type @see the fetch implementation for ExampleCas in the cas module tests

fn get_id(&self) -> Uuid