Module holochain_core_types::eav

Structs

EntityAttributeValue

ExampleEntityAttributeValueStorage

ExampleEntityAttributeValueStorageNonSync

Traits

EntityAttributeValueStorage
eav storage does NOT provide storage for AddressableContent use cas::storage::ContentAddressableStorage to store AddressableContent provides a simple and flexible interface to define relationships between AddressableContent

Functions

eav_round_trip_test_runner

test_eav

test_eav_address

test_eav_attribute

test_eav_content

test_eav_entity

test_eav_value

Type Definitions

Attribute
using String for EAV attributes (not e.g. an enum) keeps it simple and open

Entity
EAV (entity-attribute-value) data ostensibly for metadata about entries in the DHT defines relationships between AddressableContent values implemented on top of cas::storage::ContentAddressableStorage @see https://en.wikipedia.org/wiki/Entity%E2%80%93attribute%E2%80%93value_model Address of AddressableContent representing the EAV entity

Value
Address of AddressableContent representing the EAV value