Module holochain_core_types::eav::eavi

EAV stands for entity-attribute-value. It is a pattern implemented here for adding metadata about entries in the DHT, additionally being used to define relationships between AddressableContent values. See wikipedia to learn more about this pattern.

Structs

EntityAttributeValueIndex
The basic struct for EntityAttributeValue triple, implemented as AddressableContent including the necessary serialization inherited.

Enums

Attribute
All Attribute values are pre-defined here. If ever a user-defined Attribute is needed, just add a new Custom variant for it with a String parameter

AttributeError

Functions

eav_round_trip_test_runner

test_eav

test_eav_address

test_eav_attribute

test_eav_content

test_eav_entity

test_eav_value

Type Definitions

Entity
Address of AddressableContent representing the EAV entity

Index

Value
Address of AddressableContent representing the EAV value