Trait holochain_core_types::eav::EntityAttributeValueStorage

pub trait EntityAttributeValueStorage: Clone {
    fn add_eav(
        &mut self, 
        eav: &EntityAttributeValue
    ) -> Result<(), HolochainError>;

    fn fetch_eav(
        &self, 
        entity: Option<Entity>, 
        attribute: Option<Attribute>, 
        value: Option<Value>
    ) -> Result<HashSet<EntityAttributeValue>, HolochainError>;
}

eav storage does NOT provide storage for AddressableContent use cas::storage::ContentAddressableStorage to store AddressableContent provides a simple and flexible interface to define relationships between AddressableContent

Required Methods

fn add_eav(&mut self, eav: &EntityAttributeValue) -> Result<(), HolochainError>

adds the given EntityAttributeValue to the EntityAttributeValueStorage append only storage eavs are retrieved through constraint based lookups @see fetch_eav

fn fetch_eav(
&self,
entity: Option<Entity>,
attribute: Option<Attribute>,
value: Option<Value>
) -> Result<HashSet<EntityAttributeValue>, HolochainError>

fetches the set of EntityAttributeValues that match constraints None = no constraint Some(Entity) = requires the given entity (e.g. all a/v pairs for the entity) Some(Attribute) = requires the given attribute (e.g. all links) Some(Value) = requires the given value (e.g. all entities referencing an Address)