Module holochain_core_types::entry

This module contains all the necessary definitions for Entry, which broadly speaking refers to any data which will be written into the ContentAddressableStorage, or the EntityAttributeValueStorage. It defines serialization behaviour for entries. Here you can find the complete list of entry_types, and special entries, like deletion_entry and cap_entry.

Modules

cap_entries

deletion_entry

entry_type

Structs

EntryWithMeta

EntryWithMetaAndHeader

Enums

Entry
Structure holding actual data in a source chain "Item" data is stored as a JsonString

Functions

expected_entry_address
the correct address for test_entry()

expected_serialized_entry_content

test_entry
dummy entry

test_entry_a
dummy entry, same as test_entry()

test_entry_b
dummy entry, differs from test_entry()

test_entry_c

test_entry_content

test_entry_unique
dummy entry with unique string content

test_entry_value
dummy entry value

test_entry_value_a
dummy entry content, same as test_entry_value()

test_entry_value_b
dummy entry content, differs from test_entry_value()

test_entry_value_c

test_entry_with_value

test_sys_entry

test_sys_entry_address

test_sys_entry_value

test_unpublishable_entry

Type Definitions

AppEntryValue