Module holochain_core_types::entry_meta

Structs

EntryMeta
Meta represents an extended form of EAV (entity-attribute-value) data E = the entry key for hash table lookups A = the name of the meta attribute V = the value of the meta attribute txn = a unique (local to the source) monotonically increasing number that can be used for crdt/ordering @see https://papers.radixdlt.com/tempo/#logical-clocks source = the agent making the meta assertion signature = the asserting agent's signature of the meta assertion

Functions

test_attribute
dummy test attribute name

test_attribute_a
dummy test attribute name, same as test_attribute()

test_attribute_b
dummy test attribute name, differs from test_attribute()

test_meta
returns dummy meta for testing

test_meta_a
dummy meta, same as test_meta()

test_meta_b
returns dummy meta for testing against the same entry as test_meta_a

test_meta_for

test_value
dummy test attribute value

test_value_a
dummy test attribute value, same as test_value()

test_value_b
dummy test attribute value, differs from test_value()