Struct holochain_core_types::json::JsonString

pub struct JsonString(_);

track json serialization with the rust type system! JsonString wraps a string containing JSON serialized data avoid accidental double-serialization or forgetting to serialize serialize any type consistently including hard-to-reach places like Option and Result JsonString must not itself be serialized/deserialized instead, implement and use the native From trait to move between types

moving to/from String, str, JsonString and JsonString simply (un)wraps it as raw JSON data

moving to/from any other type must offer a reliable serialization/deserialization strategy

Methods

impl JsonString

pub fn null() -> JsonString

a null JSON value e.g. represents None when implementing From<Option>

pub fn empty_object() -> JsonString

pub fn is_null(&self) -> bool

pub fn to_bytes(&self) -> Vec<u8>

achieves the same outcome as serde_json::to_vec()

pub fn from_json(s: &str) -> JsonString