Struct holochain_core_types::json::JsonStringOption

pub struct JsonStringOption<T>(_);

Methods

impl<T> JsonStringOption<T>

pub fn map<__SHRINKWRAP_T, __SHRINKWRAP_F: FnMut(Option<T>) -> __SHRINKWRAP_T>(
self,
f: __SHRINKWRAP_F
) -> __SHRINKWRAP_T

Map a function over the wrapped value, consuming it in the process.

pub fn map_ref<__SHRINKWRAP_T, __SHRINKWRAP_F: FnMut(&Option<T>) -> __SHRINKWRAP_T>(
&self,
f: __SHRINKWRAP_F
) -> __SHRINKWRAP_T

Map a function over the wrapped value without consuming it.

impl<T> JsonStringOption<T>

pub fn to_option(self) -> Option<T>